Go back to old Commonlib

+++ src/papyrus/plugin.cpp
@@ -221,7 +221,7 @@ namespace {
 		if (!current3D) {
 			return; // Retry next frame
 		}
-		current3D->SetMotionType(hkpMotion::MotionType::kCharacter, true, true, true);
+		current3D->SetMotionType((uint32_t)hkpMotion::MotionType::kCharacter, true, true, true);
 		current3D->SetCollisionLayer(COL_LAYER::kCharController);
 	}
 	void DisableCollisionLayerAndMotion(StaticFunctionTag*, TESObjectREFR* ref) {
@@ -232,7 +232,7 @@ namespace {
 		if (!current3D) {
 			return; // Retry next frame
 		}
-		current3D->SetMotionType(hkpMotion::MotionType::kKeyframed, true, true, true);
+		current3D->SetMotionType((uint32_t)hkpMotion::MotionType::kKeyframed, true, true, true);
 		current3D->SetCollisionLayer(COL_LAYER::kNonCollidable);
 	}
